requirements.txt
# Python 3.11

allure-pytest==2.13.2
allure-python-commons==2.13.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.12.2
h11==0.14.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
nodeenv==1.8.0
outcome==1.2.0
packaging==23.1
platformdirs==3.5.3
pluggy==1.0.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
PySocks==1.7.1
pytest==7.3.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
virtualenv==20.23.0
wsproto==1.2.0
